merge

+++ frontend/src/app/orders/orders.component.ts
@@ -12,7 +12,7 @@ export class OrdersComponent {
   options = new RequestOptions({ withCredentials: true });
   baseUrl = 'http://localhost:8080/order/';
   orders: any;
-  selectedOrder  = {
+  selectedOrder = {
     customer: '',
     products: [{
       product: '',
@@ -30,35 +30,37 @@ export class OrdersComponent {
   };
   userData: any;
   products: any;
+  doneOrders: any;
+  activeOrders: any;
 
   constructor(public http: Http, public router: Router) {
     this.getOrders();
     this.getUsers();
     this.getProducts();
 
-   }
+  }
 
-   addRow() {
+  addRow() {
     this.newOrder.products.push({
       product: '',
       quantity: ''
     });
     console.log(this.newOrder);
-   }
+  }
 
-   addModalRow() {
-     this.selectedOrder.products.push({
-       product: '',
-       quantity: ''
-     });
-   }
+  addModalRow() {
+    this.selectedOrder.products.push({
+      product: '',
+      quantity: ''
+    });
+  }
 
-   getUsers() {
+  getUsers() {
     this.http.get('http://localhost:8080/useradmin/', this.options)
       .subscribe(data => {
         this.userData = JSON.parse(data['_body']);
         console.log(this.userData);
-  });
+      });
   }
 
   getOrders() {
@@ -68,26 +70,26 @@ export class OrdersComponent {
         if (d.err) {
           this.router.navigate(['/login']);
         } else {
-        for (let i = 0; i < d.length; i++) {
-         for (let j = 0; j < d[i].products.length; j++) {
-          if (d[i].products[j]['product'] === null) {
-            d[i].products[j]['product'] = { productname: 'Termék törölve' };
+          for (let i = 0; i < d.length; i++) {
+            for (let j = 0; j < d[i].products.length; j++) {
+              if (d[i].products[j]['product'] === null) {
+                d[i].products[j]['product'] = { productname: 'Termék törölve' };
+              }
+            }
           }
-         }
+          this.orders = d;
+          this.listActiveOrders();
+          this.listDoneOrders();
         }
-        this.orders = d;
-        console.log(this.orders);
-      }
-  });
+      });
   }
 
   getProducts() {
     this.http.get('http://localhost:8080/product', this.options)
-    .subscribe(data => {
+      .subscribe(data => {
         this.products = JSON.parse(data['_body']);
       });
   }
-
 
   editOrder() {
     this.http.put(`${this.baseUrl}${this.selectedOrder['_id']}`, this.selectedOrder, this.options)
@@ -98,27 +100,38 @@ export class OrdersComponent {
         }
       });
   }
+
   removeOrder(order) {
-    this.selectedOrder = order;
+    if (confirm('Biztos törli a rendelést?')) {
+      this.selectedOrder = order;
       this.http.delete(`${this.baseUrl}${this.selectedOrder['_id']}`, this.options)
-      .subscribe(data => {
-        console.log(data);
-        this.getOrders();
-      });
+        .subscribe(data => {
+          console.log(data);
+          this.getOrders();
+        });
+    }
   }
 
   createOrder() {
     this.newOrder.products = this.newOrder.products.filter(pr => pr.product !== '');
     this.http.post(`${this.baseUrl}`, this.newOrder, this.options)
-        .subscribe(data => {
-            console.log(data['_body']);
-            this.getOrders();
-        });
+      .subscribe(data => {
+        console.log(data['_body']);
+        this.getOrders();
+      });
   }
 
   loadModalData(order) {
     this.selectedOrder = order;
     console.log(this.selectedOrder);
+  }
+
+  listDoneOrders() {
+    this.doneOrders = this.orders.filter(order => order.status === 'done');
+  }
+
+  listActiveOrders() {
+    this.activeOrders = this.orders.filter(order => order.status === 'active');
   }
